server/requirements.txt
django==2.2.7
django-cors-headers==3.1.1
django-environ==0.4.5
django-extensions==2.2.5
django-graphql-jwt==0.2.2
graphene-django==2.6.0
psycopg2==2.8.4

# Testing

factory-boy==2.12.0
